+++ crates/sargon/src/profile/v100/entity_security_state/mod.rs
@@ -1,7 +1,9 @@
 mod entity_security_state;
 mod provisional_securified_config;
+mod provisional_securified_transaction_queued;
 mod unsecured_entity_control;
 
 pub use entity_security_state::*;
 pub use provisional_securified_config::*;
+pub use provisional_securified_transaction_queued::*;
 pub use unsecured_entity_control::*;
