+++ crates/sargon/src/profile/v100/entity_security_state/provisional_securified_config.rs
@@ -1,55 +1,13 @@
 use crate::prelude::*;
 
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, Hash)]
-pub struct ProvisionalSecurifiedTransactionQueued {
-    pub factor_instances: SecurityStructureOfFactorInstances,
-    pub txid: TransactionIntentHash,
-}
-
-impl ProvisionalSecurifiedTransactionQueued {
-    pub fn new(
-        factor_instances: SecurityStructureOfFactorInstances,
-        txid: TransactionIntentHash,
-    ) -> Self {
-        Self {
-            factor_instances,
-            txid,
-        }
-    }
-}
-impl HasSampleValues for ProvisionalSecurifiedTransactionQueued {
-    fn sample() -> Self {
-        Self::new(
-            SecurityStructureOfFactorInstances::sample(),
-            TransactionIntentHash::sample(),
-        )
-    }
-    fn sample_other() -> Self {
-        Self::new(
-            SecurityStructureOfFactorInstances::sample_other(),
-            TransactionIntentHash::sample_other(),
-        )
-    }
-}
-#[cfg(test)]
-mod provisional_securified_transaction_queued_tests {
-    use super::*;
-
-    #[allow(clippy::upper_case_acronyms)]
-    type SUT = ProvisionalSecurifiedTransactionQueued;
-
-    #[test]
-    fn equality() {
-        assert_eq!(SUT::sample(), SUT::sample());
-        assert_eq!(SUT::sample_other(), SUT::sample_other());
-    }
-
-    #[test]
-    fn inequality() {
-        assert_ne!(SUT::sample(), SUT::sample_other());
-    }
-}
-
+/// The different intermediary states of changing the security structure of an entity.
+/// This type is put in an `Option` on either `UnsecuredEntityControl` or `SecurifiedEntityControl`,
+/// and if `None` it means user has no provisionally changed security structure. If set, it contains
+/// these different variants:
+/// * `ShieldSelected` - User has selected which security shield to use for some entity,
+/// * `FactorInstancesDerived` - Sargon has provided a `SecurityStructureOfFactorInstances` but
+///     user has not made a transaction to apply it to the entity yet.
+/// * `TransactionQueued` - User has signed and queued a transaction changing to `SecurityStructureOfFactorInstances`
 #[derive(
     Serialize, Deserialize, Clone, Debug, PartialEq, Eq, Hash, EnumAsInner,
 )]
@@ -67,7 +25,7 @@ pub enum ProvisionalSecurifiedConfig {
         value: SecurityStructureOfFactorInstances,
     },
 
-    /// User has made queued a transaction to apply a `SecurityStructureOfFactorInstances`
+    /// User has signed and queued a transaction to apply a `SecurityStructureOfFactorInstances`
     /// but it has not been submitted (confirmed) yet.
     #[serde(rename = "transactionQueued")]
     TransactionQueued {
